Implement legacy secondaries
Common functionality of legacy and virtual secondaries is isolated
into ManagedSecondary class
